Fix errors from not renaming Lambdas
