+++ src/main/scala/com/gu/newsletterlistcleanse/Env.scala
@@ -1,0 +1,12 @@
+package com.gu.newsletterlistcleanse
+
+case class Env(app: String, stack: String, stage: String) {
+  override def toString: String = s"App: $app, Stack: $stack, Stage: $stage\n"
+}
+
+object Env {
+  def apply(): Env = Env(
+    Option(System.getenv("App")).getOrElse("DEV"),
+    Option(System.getenv("Stack")).getOrElse("DEV"),
+    Option(System.getenv("Stage")).getOrElse("DEV"))
+}
