+++ src/main/scala/com/gu/newsletterlistcleanse/GetCleanseListLambda.scala
@@ -8,31 +8,20 @@ import org.slf4j.{ Logger, LoggerFactory }
  * You can test your lambda by sending it the following payload:
  * {"name": "Bob"}
  */
-class LambdaInput() {
+class GetCleanseListLambdaInput() {
   var name: String = _
   def getName(): String = name
   def setName(theName: String): Unit = name = theName
 }
 
-case class Env(app: String, stack: String, stage: String) {
-  override def toString: String = s"App: $app, Stack: $stack, Stage: $stage\n"
-}
-
-object Env {
-  def apply(): Env = Env(
-    Option(System.getenv("App")).getOrElse("DEV"),
-    Option(System.getenv("Stack")).getOrElse("DEV"),
-    Option(System.getenv("Stage")).getOrElse("DEV"))
-}
-
-object Lambda {
+object GetCleanseListLambda {
 
   val logger: Logger = LoggerFactory.getLogger(this.getClass)
 
   /*
    * This is your lambda entry point
    */
-  def handler(lambdaInput: LambdaInput, context: Context): Unit = {
+  def handler(lambdaInput: GetCleanseListLambdaInput, context: Context): Unit = {
     val env = Env()
     logger.info(s"Starting $env")
     logger.info(process(lambdaInput.name, env))
@@ -44,8 +33,8 @@ object Lambda {
   def process(name: String, env: Env): String = s"Hello $name! (from ${env.app} in ${env.stack})\n"
 }
 
-object TestIt {
+object TestGetCleanseList {
   def main(args: Array[String]): Unit = {
-    println(Lambda.process(args.headOption.getOrElse("Alex"), Env()))
+    println(GetCleanseListLambda.process(args.headOption.getOrElse("Alex"), Env()))
   }
 }
